fix - security configuration, social and local user auth
