+++ src/main/java/com/bobbbaich/fb/bot/Application.java
@@ -2,7 +2,6 @@ package com.bobbbaich.fb.bot;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.security.oauth2.config.annotation.web.configuration.EnableOAuth2Client;
 
 @SpringBootApplication
 public class Application {
