+++ src/main/java/com/bobbbaich/fb/bot/config/WebSecurityConfig.java
@@ -1,0 +1,52 @@
+package com.bobbbaich.fb.bot.config;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.social.security.SpringSocialConfigurer;
+
+import javax.annotation.Resource;
+
+@Configuration
+public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
+
+    @Resource
+    private UserDetailsService userService;
+
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
+
+
+    @Override
+    protected void configure(HttpSecurity http) throws Exception {
+
+        http
+                .csrf().disable()
+                .authorizeRequests()
+                .antMatchers("/", "/signin/**").permitAll()
+                .anyRequest().authenticated();
+
+        http
+                .formLogin()
+                .loginPage("/")
+                .permitAll().and()
+                .logout()
+                .deleteCookies("JSESSIONID")
+                .permitAll()
+                .and().apply(new SpringSocialConfigurer());
+    }
+
+    @Autowired
+    @Override
+    protected void configure(AuthenticationManagerBuilder authManagerBuilder) throws Exception {
+        authManagerBuilder.userDetailsService(userService).passwordEncoder(passwordEncoder());
+    }
+}
