+++ src/main/resources/application-local.yml
@@ -7,7 +7,7 @@ spring:
   datasource:
       url: jdbc:postgresql://localhost:5432/social
       username: postgres
-      password: 123123
+      password: postgres
       driver-class-name: org.postgresql.Driver
   jpa:
     database-platform: org.hibernate.dialect.PostgreSQLDialect
