+++ src/main/resources/static/index.html
@@ -1,24 +1,22 @@
 <!doctype html>
 <html lang="en">
 <head>
-<meta charset="utf-8" />
-<meta http-equiv="X-UA-Compatible" content="IE=edge" />
-<title>Demo</title>
-<meta username="description" content="" />
-<meta username="viewport" content="width=device-width" />
-<base href="/" />
-<link rel="stylesheet" type="text/css"
-	href="/webjars/bootstrap/css/bootstrap.min.css" />
-<script type="text/javascript" src="/webjars/jquery/jquery.min.js"></script>
-<script type="text/javascript"
-	src="/webjars/bootstrap/js/bootstrap.min.js"></script>
+    <meta charset="utf-8"/>
+    <meta http-equiv="X-UA-Compatible" content="IE=edge"/>
+    <title>Demo</title>
+    <base href="/"/>
+    <link rel="stylesheet" type="text/css"
+          href="/webjars/bootstrap/css/bootstrap.min.css"/>
+    <script type="text/javascript" src="/webjars/jquery/jquery.min.js"></script>
+    <script type="text/javascript"
+            src="/webjars/bootstrap/js/bootstrap.min.js"></script>
 </head>
-<body ng-app="app" ng-controller="home as home">
-	<h1>Login</h1>
-	<div class="container" ng-show="!home.authenticated">
-		<div>
-			With Facebook: <a href="/signin/facebook">click here</a>
-		</div>
-	</div>
+<body>
+<h3>Login With Twitter</h3>
+<div>
+    <form method="post" action="/signin/twitter">
+        <p><input type="submit"></p>
+    </form>
+</div>
 </body>
 </html>
